requirements.txt
acres==0.5.0
arabic-reshaper==3.0.0
asn1crypto==1.5.1
blinker==1.9.0
Brotli==1.1.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
ci-info==0.3.0
click==8.1.8
colorama==0.4.6
configobj==5.0.9
configparser==7.2.0
cryptography==46.0.3
cssselect2==0.8.0
docx2pdf==0.1.8
et_xmlfile==2.0.0
etelemetry==0.3.1
filelock==3.19.1
Flask==3.1.2
fonttools==4.60.1
fpdf==1.7.2
freetype-py==2.5.1
html5lib==1.1
httplib2==0.31.0
idna==3.11
importlib_metadata==8.7.0
importlib_resources==6.5.2
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
looseversion==1.3.0
lxml==6.0.2
MarkupSafe==3.0.3
networkx==3.2.1
nibabel==5.3.2
nipype==1.10.0
numpy==2.0.2
openpyxl==3.1.5
oscrypto==1.3.0
packaging==25.0
pandas==2.3.3
pathlib==1.0.1
pillow==11.3.0
prov==2.1.1
puremagic==1.30
pycairo==1.28.0
pycparser==2.23
pydot==4.0.1
pydyf==0.11.0
pyHanko==0.31.0
pyhanko-certvalidator==0.29.0
PyMuPDF==1.26.5
pyparsing==3.2.5
pypdf==6.1.3
PyPDF2==3.0.1
pyphen==0.17.2
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-docx==1.2.0
pytz==2025.2
pywin32==311
pyxnat==1.6.3
PyYAML==6.0.3
rdflib==7.3.0
reportlab==4.4.4
requests==2.32.5
rlPyCairo==0.4.0
scipy==1.13.1
simplejson==3.20.2
six==1.17.0
svglib==1.6.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tqdm==4.67.1
traits==7.0.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.5.0
weasyprint==66.0
webencodings==0.5.1
Werkzeug==3.1.3
xhtml2pdf==0.2.17
zipp==3.23.0
zopfli==0.2.3.post1
gunicorn